requirements.txt
Django>=2.2.6
mysqlclient>=2.0.3
mysql-connector-python>=8.0.28
